Mockup a demo of our solar system
Not to scale, having difficulty scaling everything down correctly.

I know that if I want to display everything on one screen without panning then
I'll need to scale up the size of the planets, but I want to be able to do that
without needing to mess with the radius and orbit of the planets.