Limit rotation_angle around the star to -pi/4, pi/4
Doing this so the planets don't all look weird. I'd like to have a better way
of determining this for all of the planets together in the system, but that
might need to wait until I study some better ways of modeling systems and
the dynamics involved in creating one from, say, a dust cloud.