Reload the sidebar when you reset the system
Would actually prefer this to be done with bindings, but I think the
correct way to do that is to use the Owner and the controller should
be the owner of this view. Unfortunately, I'm not entirely sure how
that works, so I'm gimping myself and including this awful awful hack.