Fix a terrible bug with rotated orbits
So, how this basically worked was that we calculated where the
planet would be in orbit at time T around its center (the star)
X, Y by calculating Px, Py where the planet is at T, and then
translated by the amount of skew around the star that it's
orbit takes.

Now, there's one obvious bug there, in that the rotation takes
place after the point is found, so the center of rotation has
also been moved. Taking that into account, now all of the generated
planets will orbit around their stars again.

Also: color planets somewhat differently. I'd like this to be based
on some attributes of the planets themselves, but I'm happy with this
for now.