Draw habitable zone for the star
This is basically a bunch of handwaving. I spent about five hours last week
reading about habitable zones for different classes of stars and how that
is calculated, and it's basically a lot of handwaving to determine the
actual luminosity of a star given the apparent magnitude as we see it
with our telescopes here on Earth. I'm not entirely convinced that I'm
doing this 100% correctly, but I am convinced that it seems to do
the right thing in most cases, and that it presents a compelling data point
for different systems.