Planets should orbit at their respective orbit lengths
I think at one point I broke this because I was trying to get the
speed of planets orbiting down to a realistic rate, and I hadn't
noticed this was broken until Josh at NRI pointed it out to me.
Now this should be fixed, although I'm not terribly happy with how
it's scaled by a random magic number currently.