Abstract out superclass for stars, planets
Basically because I wanted to have escape velocity on both of them,
and I wanted to reuse code. Currently stars will set their mass
twice because I need to have a class level constant for that, but
I'm about to run and I wanted to make this change real quick.