Implement sidebar as a NSScrollView
Kind of interesting to encounter the issues here that you get when you're
creating view programmatically. It was difficult to debug why my content was
not visible when scrolling initially because I was creating my content
starting from the top of the screen and when this changed to be a scrollview,
it meant that most of the content was then below the bounds of the view.