Begin easing animations
This is pretty baller. Still needs one small change to figure out how to
animate towards where the planet will be when you select a planet.

Overall, I am very happy with how this turned out. NSAnimation is fairly
easy to use for custom properties (ie: not on an NSView)