Sidebar is collapsable and shows planets
Never thought it would be this difficult to create an instance of an NSScrollView
with NSViews inside that you can expand and collapse, but here you go. There
doesn't appear to be anything quite like this on cocoa pods, so I will probably
opensource this soon and see how hard it is to publish to cocoa pods.