Factor out an actor for scene logic
I'd like for the controller to manage the logic of what scene is active,
and that it would pass control to various scene actors that know how to
manage subsets of the current working environment (systems, planets,
people, civilizations, starships, asteroids, etc).

Working on a separate actor that will draw some UI for planets, but so far
all I have is a transition to fullscreen. This *should* include a fade to
black for the existing scene, or some other fancy schmancy animation.