Draw triangles more correctly
I need to research the difference between these two modes
I noticed with GL_TRIANGLE_FAN that the edges of the rects,
on the opposite end from where we start drawing, there would be
large swaths of pixels that hadn't been drawn. My best guess
would be that triangle_fan tries to fan out from the points you
draw and fill in the pixels that way, and triangle strip will
instead fill them in with strips of triangles between all of
the points you specify.