Start handling state of planet sidebar
I want each planet to be viewable separately, because there is a lot of
information associated with each. So when you expand a single view,
the rest will need to collapse, much like the current solar system view.

Trying to decide how to handle this, will probably use NSNotifications again.