Format test file.
Still don't have any tests, but I ran across this and it irked me. I very
much prefer my objective-c to look like

-($TYPE) $METHOD:($TYPE) $ARG {

}

instead of

-($TYPE)$METHOD:($TYPE)$ARG
{

}

Which just seems like a weird use of whitespace. I can see an argument for
putting the opening curly brace on a newline, but not for lacking whitespace
in the method signature.