Fix a memory warning under ARC
So, the way this has been explained to me, ARC and the compiler
make assumptions about memory management based on the name of the
method. So if your method is called initWithSomethingCool or
allocAndDoSomethingWeird then because the method starts with
alloc or init, it will assume that it returns memory that is then
managed by the callsite, because it will have received this allocated
and initialized memory.

So, for this case, because my initializer was called sol, even though
the initial call was alloc, it does not assume that this callsite
is responsible for managing the memory because the final method name
was sol and not initAsSol (or something along those lines).