Correctly clear color in between frames
Since I was clearing the buffer with glClearColor(0, 0, 0, 0),
weird things will happen with a black color that has a zero
alpha value. Apparently this would cause the graphics to be
drawn with what appeared to be inverted colors. Changing
this to be a fully opaque black color fixes this issue.