GLGEasedValue respects min/max in setCurrentValue
I forget why I didn't do this in the first place, I think I had
intended for consumers of this interface to "do the right thing"
but clearly I've changed my mind, and I think it's more correct
to enforce the minimum and maximum values.

This will be particularly useful for the next commit : controlling
the speed of time. In that case, I really want increment /
decrement with animations, but I'll just be using setValue for convenience.