Change how planets are initialized
I want the initializers for the stock "home" planets to
call the init method that was previously called initWithStar
so that we can setup things like trailers and what not. I think
that @andymoe made a good point that it's weird to do a lot
in your initializer, and that having init methods that takes
a lot of parameters is probably an antipattern.