Added an option to quit and return to main menu
There is a bug where if you are in the in-game menu (that you can reach
with the ESC key, and you "unpause" the game with spacebar, it will leave
the game in an inconsistent state. Unsure whether I want to do something
hacky or implement the pause state for being in the menu differently.

^^ This will also leak memory ^^